Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 int main() {
-    double a,b;
+    int a,b;
     char c;
-    scanf("%lf %lf",&a,&b);
+    scanf("%d %d",&a,&b);
     scanf("%c",&c);
     if(c=='+'){
-        printf("%lf",a+b);
+        printf("%d",a+b);
     }
     else if(c=='-'){
-        printf("%lf",a-b);
+        printf("%d",a-b);
     }
     else if(c=='*'){
-        printf("%lf",a*b);
+        printf("%d",a*b);
     }
     else if(c=='/'){
-        printf("%lf",a/b);
+        printf("%d",a/b);
     }
 
     return 0;
